dynsec: Rolelist as own file for consistency.
